requirements.txt
Django==5.2
djangorestframework==3.16.0
requests==2.32.3
django-filter==25.1
python-decouple==3.8